+++ src/main/java/com/natuan/citysimulator/abstracts/CityPlanAbstract.java
@@ -1,6 +1,6 @@
 package com.natuan.citysimulator.abstracts;
 
-import com.natuan.citysimulator.GUI.Building;
+import com.natuan.citysimulator.model.Building;
 import com.natuan.citysimulator.GUI.Intersection;
 import com.natuan.citysimulator.Resources.Constants;
 import com.natuan.citysimulator.interfaces.CityPlanInterface;
