+++ src/main/java/com/natuan/citysimulator/helper/MyOutput.java
@@ -4,6 +4,7 @@ import com.natuan.citysimulator.Resources.Constants;
 import com.natuan.citysimulator.model.Person;
 
 import java.io.*;
+import java.util.Scanner;
 import java.util.Vector;
 
 public class MyOutput implements Constants {
@@ -17,16 +18,20 @@ public class MyOutput implements Constants {
     private long seconds;
     private String Sdata;
     private String file_name;
+    private String file_convert_name;
     private String pop_file_name;
 
     public MyOutput() {
         MyOutput.seconds0 = System.currentTimeMillis();
+
+//        file_convert_name = getFileConvertName();
     }
 
     public MyOutput(final String file_name, final String pop_file_name) {
         MyOutput.seconds0 = System.currentTimeMillis();
         this.file_name = file_name;
         this.pop_file_name = pop_file_name;
+        this.file_convert_name = getFileConvertName();
     }
 
     public void showTime() {
@@ -72,7 +77,7 @@ public class MyOutput implements Constants {
                 final String substring3 = string3.substring(0, index3);
                 int index4 = string4.indexOf(".");
                 index4 += 2;
-                printWriter.println(this.Sdata = "" + string + ", " + substring + ", " + substring2 + "," + substring3 + "," + string4.substring(0, index4) + "     # cycle = " + n);
+                printWriter.println(this.Sdata = "" + string + ", " + substring + ", " + substring2 + "," + substring3 + "," + string4.substring(0, index4) + "," + n);
             }
             printWriter.flush();
             printWriter.close();
@@ -92,7 +97,7 @@ public class MyOutput implements Constants {
             printWriter.println(this.Sdata = "# Num People  = " + i + " moves between additions = " + j + " final Moves = " + k + "   output filename: = " + this.file_name);
             printWriter.println(this.Sdata = "# random number seed = " + lng);
             printWriter.println(this.Sdata = "# data format is: ");
-            printWriter.println(this.Sdata = "# index, time,  x,y,z \n");
+            printWriter.println(this.Sdata = "# index, time,  x,y,z ,# cycle\n");
             printWriter.flush();
             printWriter.close();
         } catch (FileNotFoundException ex2) {
@@ -100,6 +105,63 @@ public class MyOutput implements Constants {
         } catch (IOException ex) {
             ex.printStackTrace();
         }
+    }
+
+    public void writeFileDataConvert(final int pNum, final int cNum, final int i, final int j, final int k, final long lng){
+        final int numData = pNum * cNum;
+        String data[] = new String[numData];
+
+        try {
+            int header = 0;
+            FileInputStream fileInputStream = new FileInputStream(file_name);
+            Scanner scanner = new Scanner(fileInputStream);
+            final PrintWriter printWriter = new PrintWriter(new BufferedWriter(new FileWriter(this.file_convert_name)));
+            printWriter.println(this.Sdata = "# Convert File");
+            printWriter.println(this.Sdata = "# Num People  = " + i + " moves between additions = " + j + " final Moves = " + k + "   output filename: = " + file_convert_name);
+            printWriter.println(this.Sdata = "# random number seed = " + lng);
+            printWriter.println(this.Sdata = "# data format is: ");
+            printWriter.println(this.Sdata = "# index, time,  x,y,z ,# cycle\n");
+            printWriter.flush();
+
+            while (scanner.hasNextLine() && header < 5) {
+                scanner.nextLine();
+                header++;
+            }
+            int index = 0;
+            while (scanner.hasNextLine() && index < numData) {
+                data[index] = scanner.nextLine();
+                index++;
+            }
+
+            for (int pe = 0; pe < pNum; pe++) {
+                for(int ce = 0; ce < cNum; ce++){
+                    printWriter.println(data[pe + ce * pNum]);
+                }
+            }
+
+
+            printWriter.flush();
+            printWriter.close();
+            scanner.close();
+            fileInputStream.close();
+
+
+
+        } catch (FileNotFoundException ex2) {
+            System.err.println("Can't open output file");
+        } catch (IOException ex) {
+            ex.printStackTrace();
+        }
+
+
+    }
+
+    private String getFileConvertName(){
+        int index=0;
+        for(int i=0;i<this.file_name.length();i++){
+            if(this.file_name.charAt(i) == '.') index = i;
+        }
+        return new String(this.file_name.substring(0,index) + "Convert" + this.file_name.substring(index, this.file_name.length()));
     }
 }
