+++ src/main/resources/CitySimulator.properties
@@ -5,7 +5,7 @@ numPeople=1000
 conservePopulation=true
 relaxMoves=2000
 finalMoves=200
-fileName=test.txt
+fileName=test.csv
 xmlPlanFile=./src/main/resources/city_plan.xml
 dataDirectory=.
 #cityPlanClassName
